修复文件下载可能出现的bug 修复初始化缓存时无法读取配置文件的bug

+++ src/main/java/sunnn/sunsite/util/SunSiteConstant.java
@@ -19,7 +19,7 @@ public class SunSiteConstant {
 
     public static long sessionTimeout;
 
-    public static int cacheTimeout;
+    public static long cacheTimeout;
 
     public static int cacheInitCapacity;
 
@@ -63,11 +63,11 @@ public class SunSiteConstant {
         SunSiteConstant.sessionTimeout = sessionTimeout;
     }
 
-    public static int getCacheTimeout() {
+    public static long getCacheTimeout() {
         return cacheTimeout;
     }
 
-    public static void setCacheTimeout(int cacheTimeout) {
+    public static void setCacheTimeout(long cacheTimeout) {
         SunSiteConstant.cacheTimeout = cacheTimeout;
     }
